create children class to PyTorchClassifier to implement the fit method where we initialize the trainer and model objects
